functions/requirements.txt
azure-functions==1.18.0
azure-identity==1.17.1
openai==1.54.0
